Extracted out a class
Refactoring to simplify adding unit test coverage.
